1. 把"手动"模式下的通道和浇水时长都保存到 Flash 中. 2. 添加示例代码, 在"自动"模式下的间隔天数中按OK键, 会保存用户数据到    Flash. 每新保存一次会往 Flash 的下一个 Partition 中写入

+++ P201-HDMCU-Src/source/flash_manager.h
@@ -78,7 +78,7 @@ typedef struct stc_flash_manager
 {
     boolean_t bFlashEmpty;
     /*存储用户数据,用户把数据存储在此数组*/
-    uint8_t u8FlashManagerData[FLASH_MANAGER_DATA_LEN];
+    uint32_t u8FlashManagerData[FLASH_MANAGER_DATA_LEN];
     /*记录数据要保存到新Partition的首地址*/
     uint32_t u32DataStoredHeadAddr;
 } stc_flash_manager_t;
